reaching stability
Ok so i have gotten a small working example, and I know what I need to do. for tau > 1, i need to include previous values of cases. its not just the ones right before. i thought that my side filing was doing this, but i don't think it is correctly.